--- a/Algorithm.Framework/Execution/StandardDeviationExecutionModel.py
+++ b/Algorithm.Framework/Execution/StandardDeviationExecutionModel.py
@@ -63,37 +63,37 @@ class StandardDeviationExecutionModel(ExecutionModel):
            algorithm: The algorithm instance
            targets: The portfolio targets'''
         self.targetsCollection.AddRange(targets)
+        if self.targetsCollection.Count > 0:
+            for target in self.targetsCollection.OrderByMarginImpact(algorithm):
+                symbol = target.Symbol
 
-        for target in self.targetsCollection.OrderByMarginImpact(algorithm):
-            symbol = target.Symbol
+                # calculate remaining quantity to be ordered
+                unorderedQuantity = OrderSizing.GetUnorderedQuantity(algorithm, target)
 
-            # calculate remaining quantity to be ordered
-            unorderedQuantity = OrderSizing.GetUnorderedQuantity(algorithm, target)
+                # fetch our symbol data containing our STD/SMA indicators
+                data = self.symbolData.get(symbol, None)
+                if data is None: return
 
-            # fetch our symbol data containing our STD/SMA indicators
-            data = self.symbolData.get(symbol, None)
-            if data is None: return
+                # check order entry conditions
+                if data.STD.IsReady and self.PriceIsFavorable(data, unorderedQuantity):
+                    # get the maximum order size based on total order value
+                    maxOrderSize = OrderSizing.Value(data.Security, self.MaximumOrderValue)
+                    orderSize = np.min([maxOrderSize, np.abs(unorderedQuantity)])
 
-            # check order entry conditions
-            if data.STD.IsReady and self.PriceIsFavorable(data, unorderedQuantity):
-                # get the maximum order size based on total order value
-                maxOrderSize = OrderSizing.Value(data.Security, self.MaximumOrderValue)
-                orderSize = np.min([maxOrderSize, np.abs(unorderedQuantity)])
+                    remainder = orderSize % data.Security.SymbolProperties.LotSize
+                    missingForLotSize = data.Security.SymbolProperties.LotSize - remainder
+                    # if the amount we are missing for +1 lot size is 1M part of a lot size
+                    # we suppose its due to floating point error and round up
+                    # Note: this is required to avoid a diff with C# equivalent
+                    if missingForLotSize < (data.Security.SymbolProperties.LotSize / 1000000):
+                        remainder -= data.Security.SymbolProperties.LotSize
 
-                remainder = orderSize % data.Security.SymbolProperties.LotSize
-                missingForLotSize = data.Security.SymbolProperties.LotSize - remainder
-                # if the amount we are missing for +1 lot size is 1M part of a lot size
-                # we suppose its due to floating point error and round up
-                # Note: this is required to avoid a diff with C# equivalent
-                if missingForLotSize < (data.Security.SymbolProperties.LotSize / 1000000):
-                    remainder -= data.Security.SymbolProperties.LotSize
+                    # round down to even lot size
+                    orderSize -= remainder
+                    if orderSize != 0:
+                        algorithm.MarketOrder(symbol, np.sign(unorderedQuantity) * orderSize)
 
-                # round down to even lot size
-                orderSize -= remainder
-                if orderSize != 0:
-                    algorithm.MarketOrder(symbol, np.sign(unorderedQuantity) * orderSize)
-
-        self.targetsCollection.ClearFulfilled(algorithm)
+            self.targetsCollection.ClearFulfilled(algorithm)
 
 
     def OnSecuritiesChanged(self, algorithm, changes):
@@ -151,4 +151,3 @@ class SymbolData:
             for time, value in history.iteritems():
                 self.SMA.Update(time, value)
                 self.STD.Update(time, value)
-