… Engine/AlgorithmManager.cs
… Engine/Alphas/DefaultAlphaHandler.cs
… Engine/DataFeeds/DataFeedPacket.cs
… Engine/DataFeeds/SubscriptionSynchronizer.cs
… Engine/DataFeeds/TimeSliceFactory.cs
… Engine/Results/BacktestingResultHandler.cs