<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- `TimeSliceFactory` will avoid creating empty collections
- `ExecutionModels` will check target collection count before trying to
enumerate
- Reduce calls to .`TotalPortfolioValue`
- `SecurityValues` will only be created when required
- `TimeKeeper` will use TimeZone unique Id as dictionary key. The
TimeZone hash is expensive.
- `AlgorithmManager` will avoid calling `DateTime.UtcNow`,
`ConvertFromUtc()` and `RoundDownInTimeZone()`

` Performance tests C# IndicatorRibbonBenchmark: ~40% improvement`
`PR`
37.00 seconds at 21k data points per second. Processing total of 782,223 data points.
38.42 seconds at 20k data points per second. Processing total of 782,223 data points.
36.43 seconds at 21k data points per second. Processing total of 782,223 data points.
37.41 seconds at 21k data points per second. Processing total of 782,223 data points.

`master`
59.28 seconds at 13k data points per second. Processing total of 782,223 data points.
62.70 seconds at 12k data points per second. Processing total of 782,223 data points.
64.09 seconds at 12k data points per second. Processing total of 782,223 data points.
64.53 seconds at 12k data points per second. Processing total of 782,223 data points.


` Performance tests C# BasicTemplateBenchmark: ~10% improvement`
`PR`:
28.65 seconds at 58k data points per second. Processing total of 1,664,679 data points.
28.88 seconds at 58k data points per second. Processing total of 1,664,679 data points.
28.67 seconds at 58k data points per second. Processing total of 1,664,679 data points.
`master`
34.09 seconds at 49k data points per second. Processing total of 1,664,679 data points.
33.86 seconds at 49k data points per second. Processing total of 1,664,679 data points.
34.04 seconds at 49k data points per second. Processing total of 1,664,679 data points.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
None, continuous efforts for a faster Lean
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Need for speed
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->